requirements.txt
fastapi
uvicorn
pydantic
json2pdf-Converter
